@@ -46,18 +46,18 @@
     "rebuild-node-sass": "yarn add node-sass@^4.14.1"
   },
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "^7.3.7",
-    "@angular/cli": "^7.3.9",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
+    "@angular/animations": "8.0.0",
+    "@angular/cdk": "8.0.0",
+    "@angular/cli": "8.0.0",
+    "@angular/common": "8.0.0",
+    "@angular/compiler": "8.0.0",
+    "@angular/core": "8.0.0",
+    "@angular/forms": "8.0.0",
     "@angular/http": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/platform-server": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/platform-browser": "8.0.0",
+    "@angular/platform-browser-dynamic": "8.0.0",
+    "@angular/platform-server": "8.0.0",
+    "@angular/router": "8.0.0",
     "@hmcts/frontend": "^0.0.39-alpha",
     "@hmcts/nodejs-healthcheck": "^1.7.0",
     "@hmcts/properties-volume": "^0.0.13",
@@ -73,7 +73,7 @@
     "axios": "^0.21.1",
     "body-parser": "^1.19.0",
     "config": "^3.3.4",
-    "core-js": "^2.6.5",
+    "core-js": "3.1.3",
     "exception-formatter": "^2.1.2",
     "formidable": "^1.2.1",
     "glob": "^7.1.4",
@@ -116,10 +116,10 @@
     "**/**/chromedriver": "^87.0.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.10",
-    "@angular-devkit/build-ng-packagr": "^0.13.10",
+    "@angular-devkit/build-angular": "0.800.0",
+    "@angular-devkit/build-ng-packagr": "0.800.0",
     "@angular-devkit/core": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
+    "@angular/compiler-cli": "8.0.0",
     "@hmcts/eslint-config": "^1.4.0",
     "@ngrx/schematics": "^8.6.0",
     "@types/chai": "^4.2.15",
@@ -172,7 +172,7 @@
     "tsickle": "^0.35.0",
     "tslib": "^1.9.3",
     "tslint": "^5.20.1",
-    "typescript": ">=3.1.1 <3.3",
+    "typescript": "3.4.0-rc",
     "typescript-string-operations": "^1.3.1",
     "webdriver-manager": "^12.1.7"
   }
